updated feature : cleaning up foldering and finish delete add category from modal
