+++ src/pages/admin/warehouseAdmin/inputProducts/dashboard/component/style.css
